refactor: 페이지 네이밍 리팩토링 컴포넌트+Page
